Initial check-in of experimental xmonad modules.  In addition to the cabal package set-up, this includes a functioning version of the dynamic columns layout combinator, and a simple stacked layout.
This version of dynamic columns supports hierarchical stacks, window
and focus movement, and message re-broadcasting.  There are still a
few omissions and the mechanism for actions to act on the hierarchical
stack needs to be reworked so that the flat stack is only changed
through a call to XMonad.Operations.windows.


git-svn-id: http://awakening.csail.mit.edu/svn/amthrax/sys/dotfiles@1801 ee464a63-62dc-0310-8e92-97079e79a51f